research_agent/requirements.txt
fastapi>=0.115.0
pydantic>=2.7.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
requests>=2.32.0
